backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & ML
langchain==0.0.350
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langgraph==0.0.20
openai==1.3.7
anthropic==0.7.8

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Task Queue
celery==5.3.4
kombu==5.3.4

# File Processing
python-magic==0.4.27
Pillow==10.1.0

# External Services
stripe==7.8.0
sendgrid==6.10.0
twilio==8.10.0
boto3==1.34.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security & Compliance
cryptography==41.0.8
bcrypt==4.1.2
